use assembly to adding our annotations; introduce RUNTIME_ARG_CHECK_ANNOTATION to encode the number of args seperately from RUNTIME_ARG_CHECK; this is an optimization to avoid arg checking the libraries

+++ gcc_plugin/target.c
@@ -4,7 +4,7 @@
 void printk(const char*, ...);
 void f1(int);
 
-int f(int x) {
+int f(double x) {
     return -x;
 }
